Fix cpu (#27)
* fix decoders

* add declaration of dual_return_distance_threshold

* nebula_common. reorder return modes to match previous definition


---------

+++ nebula_drivers/include/hesai/decoders/pandar_64.hpp
@@ -54,7 +54,7 @@ struct Header
 
 struct Unit
 {
-  float distance;
+  double distance;
   uint16_t intensity;
 };
